+++ src/mol-gl/shader/marking/overlay.frag.ts
@@ -6,6 +6,8 @@ uniform vec2 uTexSizeInv;
 uniform sampler2D tEdgeTexture;
 uniform vec3 uHighlightEdgeColor;
 uniform vec3 uSelectEdgeColor;
+uniform float uHighlightEdgeStrength;
+uniform float uSelectEdgeStrength;
 uniform float uGhostEdgeStrength;
 uniform float uInnerEdgeFactor;
 
@@ -16,6 +18,8 @@ void main() {
         vec3 edgeColor = edgeValue.b == 1.0 ? uHighlightEdgeColor : uSelectEdgeColor;
         gl_FragColor.rgb = edgeValue.g > 0.0 ? edgeColor : edgeColor * uInnerEdgeFactor;
         gl_FragColor.a = (edgeValue.r == 1.0 ? uGhostEdgeStrength : 1.0) * edgeValue.a;
+        float edgeStrength = edgeValue.b == 1.0 ? uHighlightEdgeStrength : uSelectEdgeStrength;
+        gl_FragColor.a *= edgeStrength;
     } else {
         gl_FragColor = vec4(0.0);
     }